120 93 sorting 5 non lazy

+++ KioChess/Engine/Interfaces/ITranspositionTableService.cs
@@ -4,5 +4,5 @@ namespace Engine.Interfaces;
 
 public interface ITranspositionTableService
 {
-    TranspositionTable Create(short depth);
+    TranspositionTable Create(int depth);
 }
